refactor: split client_nf_3.rs into deposit, transfer, and withdraw modules
