+++ nightfall_client/src/drivers/rest/client_nf_3/transfer.rs
@@ -1,0 +1,481 @@
+use super::handle_client_operation;
+use crate::{
+    domain::{
+        entities::{Operation, OperationType, Transport},
+        error::TransactionHandlerError,
+        notifications::NotificationPayload,
+    },
+    get_zkp_keys,
+    initialisation::get_db_connection,
+    ports::{
+        contracts::NightfallContract,
+        db::{CommitmentDB, CommitmentEntryDB},
+    },
+    services::commitment_selection::find_usable_commitments,
+};
+use ark_bn254::Fr as Fr254;
+use ark_ec::twisted_edwards::Affine;
+use ark_ff::Zero;
+use configuration::addresses::get_addresses;
+use jf_primitives::poseidon::{FieldHasher, Poseidon};
+use lib::{
+    client_models::NF3TransferRequest,
+    commitments::{Commitment, Nullifiable},
+    contract_conversions::FrBn254,
+    get_fee_token_id,
+    hex_conversion::HexConvertible,
+    nf_client_proof::{Proof, ProvingEngine},
+    nf_token_id::to_nf_token_id_from_str,
+    plonk_prover::circuits::DOMAIN_SHARED_SALT,
+    serialization::ark_de_hex,
+    shared_entities::{Preimage, Salt},
+};
+use tracing::{debug, error, info, warn};
+use nf_curves::ed_on_bn254::{BJJTEAffine as JubJub, BabyJubjub, Fr as BJJScalar};
+use nightfall_bindings::artifacts::Nightfall;
+use serde::Deserialize;
+use ark_std::UniformRand;
+
+#[derive(Deserialize)]
+struct JubJubPubKey(#[serde(deserialize_with = "ark_de_hex")] JubJub);
+
+pub(crate) async fn handle_transfer<P, E, N>(
+    transfer_req: NF3TransferRequest,
+    id: &str,
+) -> Result<NotificationPayload, TransactionHandlerError>
+where
+    P: Proof,
+    E: ProvingEngine<P>,
+    N: NightfallContract,
+{
+    debug!("Handling transfer request: {transfer_req:?}");
+    let NF3TransferRequest {
+        erc_address,
+        token_id,
+        recipient_data,
+        fee,
+        ..
+    } = transfer_req;
+
+    // Convert the request into the relevant types.
+    let nf_token_id =
+        to_nf_token_id_from_str(erc_address.as_str(), token_id.as_str()).map_err(|e| {
+            error!(
+                "{id} Error when retrieving the Nightfall token id from the erc address and token ID {e}"
+            );
+            TransactionHandlerError::CustomError(e.to_string())
+        })?;
+    let keys = get_zkp_keys().lock().expect("Poisoned Mutex lock").clone();
+
+    let first_value = recipient_data.values.first().ok_or_else(|| {
+        error!("{id} No value provided in recipient data");
+        TransactionHandlerError::CustomError("No value provided in recipient data".into())
+    })?;
+    let value = Fr254::from_hex_string(first_value.as_str()).map_err(|e| {
+        error!("{id} Error when reading value: {e}");
+        TransactionHandlerError::CustomError(e.to_string())
+    })?;
+
+    let fee: Fr254 = Fr254::from_hex_string(fee.as_str()).map_err(|e| {
+        error!("{id} Error when reading fee: {e}");
+        TransactionHandlerError::CustomError(e.to_string())
+    })?;
+
+    let first_key = recipient_data
+        .recipient_compressed_zkp_public_keys
+        .first()
+        .ok_or_else(|| {
+            error!("{id} No recipient public key provided");
+            TransactionHandlerError::CustomError("missing recipient public key".into())
+        })?;
+
+    // Create a JSON string that represents the tuple struct content
+    let json_wrapped = format!("\"{first_key}\"");
+
+    // Note: ark_de_hex deserialization additionally ensures the point is on-curve and in correct subgroup
+    // Unit tests verify this validation behavior remains consistent
+    let deserialized_public_key: JubJubPubKey =
+        serde_json::from_str(&json_wrapped).map_err(|e| {
+            error!("{id} Could not deserialize recipient public key: {e}");
+            TransactionHandlerError::CustomError(format!(
+                "Could not deserialize recipient public key: {e}"
+            ))
+        })?;
+
+    let recipient_public_key = deserialized_public_key.0;
+
+    // Check that the recipient public key is not the identity point
+    if recipient_public_key.is_zero() {
+        error!("{id} Recipient public key cannot be the identity point");
+        return Err(TransactionHandlerError::CustomError(
+            "Recipient public key cannot be the identity point".to_string(),
+        ));
+    }
+
+    let ephemeral_private_key = {
+        // thread_rng() is already thread-local and cheap to create; no need to share via RwLock.
+        let mut rng = ark_std::rand::thread_rng();
+        BJJScalar::rand(&mut rng)
+    };
+    let shared_secret: Affine<BabyJubjub> = (recipient_public_key * ephemeral_private_key).into();
+
+    // add the id to the request database
+
+    // Select the commitments to be spent.
+    let spend_commitments;
+    {
+        let db = get_db_connection().await;
+        let fee_token_id = get_fee_token_id();
+        let spend_value_commitments = find_usable_commitments(nf_token_id, value,db)
+        .await.map_err(|e|{
+            error!("{id} Could not find enough usable value commitments to complete this transfer, suggest depositing more tokens: {e}");
+            TransactionHandlerError::CustomError(e.to_string())})?;
+        let spend_fee_commitments = if fee.is_zero() {
+            [Preimage::default(), Preimage::default()]
+        } else {
+            match find_usable_commitments(fee_token_id, fee, db).await {
+                Ok(commitments) => commitments,
+                Err(e) => {
+                    debug!("{id} Could not find enough usable fee commitments, suggest depositing more fee: {e}");
+                    // rollback the value commitments to unspent if fails to find fee commitments
+                    let value_commitment_ids = spend_value_commitments
+                        .iter()
+                        .filter_map(|c| match c.hash() {
+                            Ok(h) => Some(h),
+                            Err(e) => {
+                                error!("{id} Failed to hash commitment during rollback, commitment may be orphaned: {e}");
+                                None
+                            }
+                        })
+                        .collect::<Vec<_>>();
+
+                    for commitment_id in &value_commitment_ids {
+                        if let Some(existing) = db.get_commitment(commitment_id).await {
+                            let _ = db
+                                .mark_commitments_unspent(
+                                    &[*commitment_id],
+                                    existing.layer_1_transaction_hash,
+                                    existing.layer_2_block_number,
+                                )
+                                .await;
+                        }
+                    }
+                    return Err(TransactionHandlerError::CustomError(e.to_string()));
+                }
+            }
+        };
+        spend_commitments = [
+            spend_value_commitments[0],
+            spend_value_commitments[1],
+            spend_fee_commitments[0],
+            spend_fee_commitments[1],
+        ];
+    }
+
+    // Work out how much change is needed.
+    let total_token_value = spend_commitments[..2]
+        .iter()
+        .map(|c| c.get_value())
+        .sum::<Fr254>();
+
+    let token_change = total_token_value - value;
+    let total_fee_value = spend_commitments[2..]
+        .iter()
+        .map(|c| c.get_value())
+        .sum::<Fr254>();
+    let fee_change = total_fee_value - fee;
+
+    let poseidon = Poseidon::<Fr254>::new();
+    // Derive a shared salt from the shared secret using domain-separated Poseidon hash.
+    let shared_salt_hash = poseidon
+        .hash(&[shared_secret.x, shared_secret.y, DOMAIN_SHARED_SALT])
+        .map_err(|e| {
+            error!("{id} Failed to derive shared salt with Poseidon: {e}");
+            TransactionHandlerError::CustomError(e.to_string())
+        })?;
+    let shared_salt = Salt::Transfer(shared_salt_hash);
+
+    // transferred value commitment, salt is derived from the shared secret
+    let new_commitment_one = Preimage::new(
+        value,
+        nf_token_id,
+        spend_commitments[0].get_nf_slot_id(),
+        recipient_public_key,
+        shared_salt,
+    );
+
+    let new_commitment_two = if !token_change.is_zero() {
+        Preimage::new(
+            token_change,
+            nf_token_id,
+            spend_commitments[0].get_nf_slot_id(),
+            keys.zkp_public_key,
+            Salt::new_transfer_salt(),
+        )
+    } else {
+        Preimage::default()
+    };
+
+    let nightfall_address = FrBn254::from(get_addresses().nightfall()).0;
+    let contract_nf_address = Affine::<BabyJubjub>::new_unchecked(Fr254::zero(), nightfall_address);
+
+    let fee_token_id = get_fee_token_id();
+    // if fee is zero, then no fee commitment is needed
+    let new_commitment_three = if !fee.is_zero() {
+        Preimage::new(
+            fee,
+            fee_token_id,
+            fee_token_id,
+            contract_nf_address,
+            Salt::new_transfer_salt(),
+        )
+    } else {
+        Preimage::default()
+    };
+
+    let new_commitment_four = if !fee_change.is_zero() {
+        Preimage::new(
+            fee_change,
+            fee_token_id,
+            fee_token_id,
+            keys.zkp_public_key,
+            Salt::new_transfer_salt(),
+        )
+    } else {
+        Preimage::default()
+    };
+
+    let new_commitments = [
+        new_commitment_one,
+        new_commitment_two,
+        new_commitment_three,
+        new_commitment_four,
+    ];
+
+    debug!(
+        "{id} New commitment hashes: {:?}",
+        new_commitments
+            .iter()
+            .filter_map(|c| match c.hash() {
+                Ok(h) => Some(h.to_hex_string()),
+                Err(e) => {
+                    warn!("{id} Failed to hash commitment for debug output: {e}");
+                    None
+                }
+            })
+            .collect::<Vec<_>>()
+    );
+
+    let secret_preimages = [
+        spend_commitments[0].get_secret_preimage(),
+        spend_commitments[1].get_secret_preimage(),
+        spend_commitments[2].get_secret_preimage(),
+        spend_commitments[3].get_secret_preimage(),
+    ];
+    let op = Operation {
+        transport: Transport::OffChain,
+        operation_type: OperationType::Transfer,
+    };
+    match handle_client_operation::<P, E, N>(
+        op,
+        spend_commitments,
+        new_commitments,
+        ephemeral_private_key,
+        Fr254::zero(),
+        secret_preimages,
+        id,
+    )
+    .await
+    {
+        Ok(res) => Ok(res),
+        Err(e) => {
+            //  rollback to UNSPENT status if handle_client_operation fails
+            let db = get_db_connection().await;
+
+            // Rollback the spend commitments to unspent
+            let commitment_ids = spend_commitments
+                .iter()
+                .filter_map(|c| match c.hash() {
+                    Ok(h) => Some(h),
+                    Err(e) => {
+                        error!("{id} Failed to hash commitment during rollback, commitment may be orphaned: {e}");
+                        None
+                    }
+                })
+                .collect::<Vec<_>>();
+
+            info!(
+                "{id} Rolling back {} spend commitments",
+                commitment_ids.len()
+            );
+
+            for commitment_id in &commitment_ids {
+                if let Some(existing) = db.get_commitment(commitment_id).await {
+                    let _ = db
+                        .mark_commitments_unspent(
+                            &[*commitment_id],
+                            existing.layer_1_transaction_hash,
+                            existing.layer_2_block_number,
+                        )
+                        .await;
+                }
+            }
+            // Delete new commitments
+            let new_commitment_ids = new_commitments
+                .iter()
+                .filter_map(|c| match c.hash() {
+                    Ok(h) => Some(h),
+                    Err(e) => {
+                        error!("{id} Failed to hash commitment during rollback, commitment may be orphaned: {e}");
+                        None
+                    }
+                })
+                .collect::<Vec<_>>();
+
+            info!("{id} Deleting {} new commitments", new_commitment_ids.len());
+            let _ = db.delete_commitments(new_commitment_ids).await;
+
+            Err(TransactionHandlerError::CustomError(e.to_string()))
+        }
+    }
+}
+
+#[cfg(test)]
+mod tests {
+    use super::*;
+    use ark_ff::One;
+    use ark_serialize::{CanonicalSerialize, Compress};
+    use ark_std::Zero;
+    use lib::{
+        client_models::NF3RecipientData,
+        plonk_prover::plonk_proof::{PlonkProof, PlonkProvingEngine},
+    };
+    use nf_curves::ed_on_bn254::BabyJubjub;
+    use nf_curves::ed_on_bn254::Fq;
+
+    /// Tests that transfer API rejects invalid recipient public keys
+    #[tokio::test]
+    async fn test_transfer_api_rejects_invalid_recipient_keys() {
+        // Invalid compressed point (not on the curve) should fail early in handle_transfer
+        let invalid_transfer_req = NF3TransferRequest {
+            erc_address: "0x1234567890123456789012345678901234567890".to_string(),
+            token_id: "0x00".to_string(),
+            recipient_data: NF3RecipientData {
+                values: vec!["0x04".to_string()],
+                recipient_compressed_zkp_public_keys: vec![
+                    "0x000ffffffffffffffffffffffffffffffffffffffffffffffffffffffffff000"
+                        .to_string(),
+                ],
+            },
+            fee: "0x00".to_string(),
+        };
+
+        let result = handle_transfer::<PlonkProof, PlonkProvingEngine, Nightfall::NightfallCalls>(
+            invalid_transfer_req,
+            "test-id-1",
+        )
+        .await;
+
+        // This should fail at the recipient key validation stage, demonstrating the API validates keys
+        assert!(
+            result.is_err(),
+            "Transfer API should reject invalid recipient public key"
+        );
+        if let Err(TransactionHandlerError::CustomError(msg)) = result {
+            assert!(
+                msg.contains("Could not deserialize recipient public key: the input buffer contained invalid data"),
+                "Error should indicate recipient public key deserialization failure, got: {msg}"
+            );
+        } else {
+            panic!("Expected TransactionHandlerError::CustomError with recipient public key deserialization failure");
+        }
+    }
+
+    #[tokio::test]
+    async fn test_transfer_api_rejects_identity_recipient_keys() {
+        // Identity point should fail early in handle_transfer
+        let identity_point = Affine::<BabyJubjub>::zero();
+        let mut compressed_bytes = Vec::new();
+        identity_point
+            .serialize_with_mode(&mut compressed_bytes, Compress::Yes)
+            .unwrap();
+        compressed_bytes.reverse(); // Convert to big-endian to match ark_se_hex format
+        let identity_point_hex = format!("0x{}", hex::encode(compressed_bytes));
+
+        let identity_point_transfer_req = NF3TransferRequest {
+            erc_address: "0x1234567890123456789012345678901234567890".to_string(),
+            token_id: "0x00".to_string(),
+            recipient_data: NF3RecipientData {
+                values: vec!["0x04".to_string()],
+                recipient_compressed_zkp_public_keys: vec![identity_point_hex],
+            },
+            fee: "0x00".to_string(),
+        };
+
+        let result = handle_transfer::<PlonkProof, PlonkProvingEngine, Nightfall::NightfallCalls>(
+            identity_point_transfer_req,
+            "test-id-2",
+        )
+        .await;
+
+        assert!(
+            result.is_err(),
+            "Transfer API should reject recipient public key if it is the identity"
+        );
+        if let Err(TransactionHandlerError::CustomError(msg)) = result {
+            assert!(
+                msg.contains("Recipient public key cannot be the identity point"),
+                "Error should indicate recipient public key cannot be the identity point, got: {msg}"
+            );
+        } else {
+            panic!("Expected TransactionHandlerError::CustomError with recipient public key cannot be the identity point");
+        }
+    }
+
+    #[tokio::test]
+    async fn test_transfer_api_rejects_low_order_recipient_keys() {
+        // A point that is low order but on the curve should fail early in handle_transfer
+        // We use point (0, -1) which is order 2 on BabyJubJub
+        let zero_x = Fq::zero();
+        let neg_one_y = -Fq::one();
+
+        let low_order_point = Affine::<BabyJubjub>::new_unchecked(zero_x, neg_one_y);
+
+        let mut compressed_bytes = Vec::new();
+        low_order_point
+            .serialize_with_mode(&mut compressed_bytes, Compress::Yes)
+            .unwrap();
+        compressed_bytes.reverse(); // Convert to big-endian to match ark_se_hex format
+        let low_order_hex = format!("0x{}", hex::encode(compressed_bytes));
+
+        let low_order_transfer_req = NF3TransferRequest {
+            erc_address: "0x1234567890123456789012345678901234567890".to_string(),
+            token_id: "0x00".to_string(),
+            recipient_data: NF3RecipientData {
+                values: vec!["0x04".to_string()],
+                recipient_compressed_zkp_public_keys: vec![low_order_hex],
+            },
+            fee: "0x00".to_string(),
+        };
+
+        let result = handle_transfer::<PlonkProof, PlonkProvingEngine, Nightfall::NightfallCalls>(
+            low_order_transfer_req,
+            "test-id-3",
+        )
+        .await;
+
+        // This should fail at the explicit .check() stage since zero point is low-order
+        assert!(
+            result.is_err(),
+            "Transfer API should reject low-order recipient public key"
+        );
+        if let Err(TransactionHandlerError::CustomError(msg)) = result {
+            assert!(
+                msg.contains("Could not deserialize recipient public key: the input buffer contained invalid data"),
+                "Error should indicate recipient public key deserialization failure, got: {msg}"
+            );
+        } else {
+            panic!("Expected TransactionHandlerError::CustomError with recipient public key deserialization failure");
+        }
+    }
+}
